requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.108.0
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.2
MarkupSafe==2.1.3
orjson==3.9.10
passlib==1.7.4
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.32.0.post1
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
